requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
nltk>=3.8.0
sumy>=0.11.0
google-generativeai>=0.3.0
click>=8.1.0
